+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Remove model asset column (#636)
+- Remove performance asset column (#640)
 
 ## [0.9.2] - 2022-04-15